move helpers out of geh_stream
