+++ source/databricks/tests/integration/calculator/steps/wholesale/test_tariff_calculators.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 from decimal import Decimal
 from datetime import datetime
-from tests.geh_stream.helpers.test_schemas import (
+from tests.helpers.test_schemas import (
     tariff_schema,
     tariff_sum_and_count_schema,
     tariff_distinct_schema,
