+++ source/databricks/tests/integration/calculator/steps/wholesale/test_wholesale_initializer.py
@@ -40,7 +40,7 @@ from geh_stream.schemas import (
     market_roles_schema,
 )
 from geh_stream.schemas import time_series_points_schema
-from tests.geh_stream.helpers.test_schemas import (
+from tests.helpers.test_schemas import (
     charges_with_prices_schema,
     charges_with_price_and_links_schema,
     charges_with_price_and_links_and_market_roles_schema,
